Use lfp as the underlying io-device for dl:stream
In an effort to support Tape Image Format (TIF) files, the realisation
that the amount of bookkeeping needed to keep track of _two_ access
mechanic format, Visible Envelope (VE) and TIF, is just to large for
dlisio to tackle on its own. Luckily the library Layered File Protocols
[1] is designed for this kind of bookkeeping, and more precisely;
present the illusion that these access mechanic formats are not there.
From lfp's documentation:

 > Layered file protocols is a high-level interface for byte-oriented
   file handles with layering. File descriptors are composed and stacked
   at run-time, adding features such as pre-fetching, indices, cloud
   access, and file format support at each level

lfp have out of the box support for both Visible Envelope and Tape
Image Format. These protocols give the user of the io-device the
illusion that the file is without Visible Envelope and tapemarks, respectably.
For dlisio, that means that our core library can act as if the only
thing present in a dlis-file is the actual dlis-data.

This commit only adapts dlisio's main io-device, dl::stream, to use
lfp's cfile is favor of std::fstream. The real golden nugget comes later
when lfp's rp66 and tapeimage protocols are introduced.

[1] https://github.com/equinor/layered-file-protocols

fixup! Use lfp as the underlying io-device for dl:stream