Add support for files encapsulated in TIF
With lfp as the only io-device, adding support for TapeImageFormat-files
is trivial. In the same way as the rp66-protocol hides all bytes that
are part of the Visible Envelope, lfp's tapeimage protocol hides all
bytes that are apart of the Tape Image Format:

What the physical file look like on disk:

               ------------------------------------------
              | TM | SUL | TM | VR | LRS | .. | LRS | TM |
               ------------------------------------------
byteoffsets:  0   12    92   104  108   208  308  408  420

What dlisio see after applying both the tapeimage- and rp66-protocol
from lfp:

               -----------------
              | LRS |  .. | LRS |
               -----------------
byteoffsets:  0   100    200   300

dlisio investigates the first bytes of a file to figure out
whether or not there are tapemarks present. Protocols are added
accordingly. In effect, dlisio does not put any responsibility on the
user to supply information about the presence of tapemarks.

There is currently no support for files with missing storage units or
missing tapemarks at the end. Both are clear standard violations.
However, missing suls seams to be fairly common (~10% of TIF files), and
the logic for such support should be fairly simple, but for now this is
considered a low hanging fruit for the near future.