Fix Visible Envelope's Format Verson in testfile
The Format Version (FV) is changed from 0x02 to 0x01 which is the
correct value for a DLISv1 file. Currently dlisio does not care about
the value of FV, but in the foreseeable future that might change. As the
purpose of this testfile is unrelated to the FV, we correct it in order
to avoid failure on FV:

  printf '\x01' | dd of=3lr-in-vr-one-encrypted.dlis bs=1 seek=83 count=1
  conv=notrunc