Fix too long VRL in testfile
The Visible Record Length in this file was longer than the actuall VR.
I.e. the VR was truncated. It's an intresting find, as dlisio apparently
does _not_ notice the truncation when it aligns perfectly with the end
of Logical Record. This is subject to change when layered file protocols
is introduced - so for now just update the VRL in the testfile:

printf '\x16' | dd of=multidimensions-multifdata.dlis bs=1 seek=81
count=1 conv=notrunc