Remove tif-ish garbage bytes from testfiles
Before the knowledge of Tape Image Format, we considered tapemarks as
random trash bytes. Although not supporting TIF directly, dlisio is able
to deal with such unknown bytes at the start of the file. The two
test files for this behaviour have trash bytes that are valid tapemarks.
This is not a problem now, but when TIF support is added we would still
like to be able to test support for actual trash bytes. Hence the
tif-like trash in these test files is changed to actual trash:

data/chap2/pre-sul-garbage.dlis

  for i in {0..11}; do printf '\xaa'; done | dd of=pre-sul-garbage.dlis
  bs=12 count=1 conv=notrunc

data/chap2/pre-sul-pre-vrl-garbage.dlis

  for in {0..11}; do printf '\xaa'; done |
  dd of=pre-sul-pre-vrl-garbage.dlis bs=12 count=1 conv=notrunc
  &&
  for i in {0..23}; do printf '\xaa'; done |
  dd of=pre-sul-pre-vrl-garbage.dlis bs=1 count=24 seek=92 conv=notrunc

Additionally, the VRL of the last Visible Record was to long. Currently
dlisio does care as long as it where able to read what seems to be all
the data. However, this is considered a bug and will change in the
future. The VRL is updated to match the actual size:

  printf '\x0a\x84' | dd of=pre-sul-garbage.dlis bs=1
  seek=4592 count=2 conv=notrunc

  printf '\x0a\x84' | dd of=pre-sul-pre-vrl-garbage.dlis bs=1
  seek=4616 count=2 conv=notrunc