Remove duplicated doc entries for some properties
These properties were documented first in the class-docstring under the
section Attributes, and then again with a property docstring. According
to the numpy styleguide [1] which we try to follow, this is good
practice. However, it turns out that prior to sphinx v3.0.0 these are
rendered twice, and from v3.0.0 an warning is issued. We deliberately
fail out CI-pipeline on any warnings, hence we caught this now.

[1] https://numpydoc.readthedocs.io/en/latest/format.html#class-docstring