Check 'instance' in IndexedObjectDescriptor
The instance argument refers to the instance of which the attribute was
accessed through. However, should the attribute be accessed directly
from the class like so

 dlisio.dlis.channels

then 'instance' is defined to be None. That means our implementation of
__get__ fails with 'TypeError: 'NoneType' object is not subscriptable'.

Although making calls that would invoke this behavior makes little
sense as a end-user, sphinx' autodoc directive makes all sorts of calls
in order to build docs and examples.