docs: Re-write the curves-section
An integral part in efficiently accessing and working with curve-data is
to understand the Frame--Channel relationship and know about the many
pitfalls and none-trivialities that lurks. The new curve-section tries
to enlighten the readers on these subjects and give them the tools
needed to work not only the trivial cases, but the more complex ones as
well.