Use fingerprint > name in Channel.curves
Indexing by name only works if the mnemonic is unique within the frame,
while indexing with fingerprint always works.

Note that Channel.curves does not, or did not work for frames with
duplicated Channels (with the same fingerprint). This is a clear
standard violation so that is considers to be ok.