Install lfp on CircleCI nodes
This patch does not add lfp to dlisio in any way, it only installs lfp
in the test containers to prime them for when lfp is included into
dlisio.

lfp has no pre-built binaries yeti, so build from source. lfp assumes
fmtlib is installed on the system as well, so install that too [1].
Building from source (especially fmtlib) will increase the total runtime
of each node a bit. But that will improve in the future - when pre-built
binaries for lfp is made available.

[1] the debian package libfmt-dev does not export the target
fmt::fmt-header-only which is the one we want, so install from source.