Add preprocessor definitions for endianness
Test endianness with cmake's TestBigEndian and set preprocessor
definitions HOST_BIG_ENDIAN and HOST_LITTLE_ENDIAN. The missing
definition of HOST_BIG_ENDIAN was a bug, as we have code that assumes it
to be defined on big endian systems. This was never caught as we
currently do not test on any big endian machines.

Although not currently used, add HOST_LITTLE_ENDIAN for completeness.