Operate on local data copy in dlis_tapemark
For a time being we need to specially deal with tapemarks in dlisio.

But when the system is big-endian, we attempted to change const buffer
passed to us from outside, which results in error.
As we still need to process data, but do not want to spoil them for
anyone else, create the local copy of provided data for big-endian
systems.