Close all filehandles on error in dlisio.load
If the flow fails, we do not return Batch object, so we leak filehandles
from previous logical files. We need to prevent that.
Additionally, last filehandle needs to be closed explicitly as it's not
on the list of correctly loaded logical files.

Depending on the file structure there are many various options at where
load can fail (and not all of them are tested), so we have to assure we
release resources in any case.