Attach object type to basic_object
The type of an object is defined on the object set, not on the actual
object, thus an instances of basic_object must remain linked to it's
set in order to have a full representation of the object.

The concept of a object set is not semantically intresting after
the objects in the set are parsed. But because of the type, the set has
to be carried through all the way to python.

By setting the type directly on the basic_object when it's parsed the
need to keep it close to the set goes away. It also fasilitates the
posibility of removing object sets from the C++ API all together.