Formalize and enrich the index
Logical Record Segment Headers contain useful information when
determining the content of the record. Previously we had to extract the
entire record in order to retreive information about it's content -
often requiring our control flow to extract more records than needed.

The lrsh are already being read when indexing so the only cost here is
an increase in the size of the index.