Pass down generator platform for VS
When attempting to setup CI/CD pipeline for different provider on
Windows, the following issue occurred:
when we attempt to build wheels for 32-bit Windows on 64-bit Windows
machine, build errors (error LNK2019, warning LNK4272 "library machine
type 'x86' conflicts with target machine type 'x64'")

This happens as we do not pass platform options down to skbuild.

If our build host and build target are the same, we do not have any
issues.

But, for example, to build 32-bit version on 64-bit linux we need to
pass -DCMAKE_CXX_FLAGS="-m32" everywhere. (For releases we build with
manylinux, so we do not need to do this explicitly).

For mcvs option however is -A Win32, which is supported only by certain
generators [1]

Giving the variation of build configurations, supporting all of them is
not straightforward.
Hence this commit means to add support for just one configuration:
building on 64-bit/32-bit host system for 32-bit/64-bit target system
with Visual Studio generators.

[1]
https://cmake.org/cmake/help/latest/variable/CMAKE_GENERATOR_PLATFORM.html