Use dl:basic_object as attic for Python types
This is a small step in a larger quest of moving the pool of metadata
objects to c++ which will give a performance boost.

There is a nice side-effect to this change: Because attic is now a c++
type, the values of object-attributes are not casted to python types
until the attribute is access. Thus UnicodeErrors that were previously
emitted on object construction is now postpone until the attribute is
accessed.