Move the channel -> frame referencing out of load
Creating a back-reference from channels to their parent frame is rather
awkward. Ideally we would completely remove it, but that would require
some notable changes to the api. So it stays for now.

The c++ api doesn't currently take part in these back-references between
objects - and we want it to stay that way. The loading routine will
gradually be moved to c++, hence the referencing is moved so that the
porting doesn't require c++ to do back-referencing.