Make dl::object_set a self-parsing type
Experience with the python implementation of parsing objects
on-demand has shown that the complexity of bookkeeping the state of
each object-set quickly becomes complicated.

To combat some of that complexity, the object-parsing is moved to the
set itself. The state of the set (i.e. parsed or unparsed) then becomes
an implementation detail of the set. Thus freeing any logical file
implementation of bookkeeping the state of all its object-sets.