Add a queryable pool of metadata object
dl::pool is intended to relieve python's dlis object of the
responsibility of caching raw record, parsed objects and the tedious
task of parsing records at the correct time.

The c++ counterpart is conceptually simpler as each object_set
are responsible for parsing. That means dl::pool is a simple
container with some convenient functions for querying objects.