Replace on-fly object creation in tests with files
Creation of objects on fly isn't so easy anymore as we need to supply
more and more values, and implementation keeps changing.
Hence rewriting some tests to be based on actual files instead.

This commit doesn't touch however on-fly tests which refered to singular
objects and didn't need logical_file property set as for now they don't
seem to be affected.