Pass logical file directly as lookup parameter
For lookup it's not actually interesting to know which object orders the
lookup. We need barely the reference to logical file itself (which
contains object pool).

With the idea of moving tests to files and stopping their on-fly
creation, situation where object will not have a logical file assigned
is not supposed to happen anymore.