Set infocodes on dl::object_attribute when parsing
Due to the intricate structure of a dlis-file, dlisio over-parse when a
certain pieces of information is queried. This would typically makes any
warnings or errors pre-mature and might result in the query failing due
to an error in some (from a user-perspective) unrelated data.

To combat this issue, any errors, inconsistencies or actions taken by
parse_object is communicated through enums set on the parsed objects.
The consumer of the objects are then responsible for investigating the
this information before using the object.

The intention is to follow up on with the same approach basic_object and
object_set.