Specifically handle "path is directory" situation
Providing incomplete path (directory only) is quite an easy mistake
during the use of dlisio.

There is no standard for fopen when file is a directory. Some
implementations did not fail on that, which resulted in a very
unintuitive error message.

LFP was updated to have some control over it and throw a custom message
on the systems where fopen succeeds and fread fails.

However it appears that for systems where fopen does succeed, error
message still might be quite confusing.

Hence decision was made to specifically check in python that path is an
existing file. Additional to the path is directory situation, check will
take over non existing paths.
All other possible ioerror cases would continue be handled by LFP.