Fix dl::value_vector comparison for monostate
It the current type of an value_vector is mpark::monostate, mpark::visit
picks the rvalue overload in variant_equal, which makes sense. However
it also picks that overload when both lhs and rhs are monostate,
resulting in a false negative for the comparison.

A new overload with the same typename for both arguments would also
resolve the problem, but I'm afraid that it would swallow more than it
should.