Make std::vector< dl::index_entry > an opaque type
Because the memory layout of python lists and std::vector differs,
pybind needs to do a copy operation during conversion.

The index is passed back and forth between c++ and python, which results
in at least two copies during load(). As the content of the index is
never actually examined in python, these copies are completely
unnecessary - and costly as the index may be quite large.

Rendering them opaque means they can be passed by reference by pybind.