Add CircleCI big-endian job
When deciding which CI to use for bigendian check the following criteria
were considered:
1. On travis we are able to build on native machines, on CircleCi - only
with the help of qemu.
2. CircleCI build time is very long, especially if docker images are not
stored and maintained. Travis is way faster.
3. Travis multi-architecture solution didn't seem to be very stable.

Travis won.

But now it seems like Travis bigendian job fails more often than it
passes (due to reasons independent from dlisio).
Hence adding CircleCI solution, given that it was actually implemented
already.
We shall see which of the jobs frustrates us less.