Implement a python version of exact_matcher
Doing the string comparison in python is more robust against different
encodings than doing the same comparison in c++.

Assuming there exist an object with name 'КАДР' and we try to access it
with f.object like so:

        f.object('TYPE', 'КАДР')

If the comparison happen in c++, the encoded 'КАДР' string is cast'ed to
a std::string and compared to objects in c++. Due to the way the
conversion for encoded python strings to std::string works, no match is
found. The same is true for byte objects, which is the default fall back
for dlisio's decode scheme.

When the comparison takes place in python, the object name as read
from file are cast'ed from dl::ident to python string. This cast runs
through our encoding schema and will produce a python string that is a
match with the parameter as long as the correct encoding is specified
with dlisio.set_encodings. If no encoding is specified the object can
still be reached with the byte object:

        f.object('TYPE', b'\xeb\xe1\xe4\xf2')

Although it's unlikely that users will know the correct byte pattern to
make that query, it's very useful internally in dlisio. This basically
assures us that internal object lookups are unaffected by encodings and
will always work.

The performance hit should not be too bad. dl::pool::get does a two-face
lookup: first on type, then on name. Thus the number of comparisons for
a single call is generally low.