Return dl::ident from fingerprint functions
This change is necessary because of a shortcoming in our string
decoding. String decoding takes place during the cast from a c++ type to
a python string. For our custom types we have custom type-casters where
we can apply string encoding as we like. But std::string's are casted
automatically by pybind's caster, which we do not have the same control
over. Hence returning non-ascii std::string's result in an UnicodeError.

Pybind11 does support encoding strings, but it requires an explicit
cast. That means every function returning std::strings has to be
aware of the available encodings and act accordingly. That is
error-prune compared to letting the cast do the encoding.