Return a std::map from findfdata
This is a performance thing only. A typical file can have thousands of
fdata records but just a handful of Frames. Using a map means we only
have to cast one string pr Frame, instead of one for each fdata record.

On my machine, the execution time for creating the fdata_index was
reduced with about 5 percent. The file contained ~30k fdata records, all
belonging to the same Frame.