Immediately store procesed data in object set
With current approach if error happens during object set processing, we
lose all the processed template attributes and objects.

Storing them the moment we processed them seems like a better approach.
Thus if error happens afterwards we are in better situation with regards
to debugging, error recovery and feeding users as much data as possible.