Introduce dlis_error and error_severity
dlis_error and error_severity are two beings that we would need later
when creating our error-mechanism.

dlis_error represents error one way or another related to protocol
processing.
error_severity is a field in dlis error which shows how critical for
further and correct processing of data is this error according to
developers.