Get rid of creating tests through attic
For convenience we still created some tests by changing attic, not
through files.
This however made it hard to change implementation without changing all
the tests first. We also cheated for a while with creating dictionaries
in place of attic.

Once basic_object would return object_attribute, not just the value,
this no longer would be the simple solution and would be a burden.

Hence files are created instead.