Completely drop support for attributes change
Decision was long in the making, but we no longer support any changes in
the attributes.
Users can't change or delete existing values anymore even using back
doors.

dlisio has no write-back support. This feature existed before to allow
users to change current object in memory as per their wishes.
Feature was mostly driven by dynamic nature of python language.

Now as we moved more implementation to c++, feature naturally
disappeared. It would require effort to re-implement it in c++
basic_object. Currently there seems little need for it, so it's not on
the road map.