Simplify init of BasicObject + derived classes
The reasoning for including 'name' and 'type' as defaulted arguments in
the BasicObject constructor is mostly historic, and some legacy tests
used them to construct 'dummy' objects for specific test-cases. That is
no longer the case as of [1], thus there is really no good reason to
keep them. 'lf' had a default value for the same reason. In real-world
code a BasicObject or a derived version is pointless without a reference
to it's logical file, hence it's removed too.

[1] SHA: 10215cf64fbaaadfb52b49c11522adb3821e75a5