Use 'find' as backend for 'match'
Although deprecated, change the internals of match to a call to 'find'
in order to avoid having to maintain the implementation of match. This
also means we that we are not dependent on removing match in order to
fulfill the goal of having one single entry point to the c++ pool,
through find.