Deprecate logicalfile.__getitem__
Schedule logicalfile.__getitem for removal. It's existence was
justified because match and object was unsuited for type-based queries.
With 'find' that is no longer the case and the need for __getitem__ is
gone.

Although it's a part of the API, __getitem__ has been living in the
shadow of functions like f.channels, f.frames etc - and using it
directly has never been encouraged.