Add error_handler
As we plan to finally give users all the warnings we see during
operations in C++, we need to create an error handler class familiar
both to C++ and to Python.
We'll do it similar way to the PyMatcher.

While logger conceptually could be a global property (like encodings),
decision was made to avoid global settings and make logger a dlis
attribute.
It's likely to stay the same for various files, but now user is in
control.