Inform objects about problems in attributes
When something unexpected happens in one of the attributes, we
want to let users know about it when they access other attributes of the
same object.

Due to small amount of restrictions in rp66 it can happen that we
correctly process already broken attributes and do not notice that they
are spoiled. Thus user will trust these attributes when they shouldn't.

In theory same can happen for objects (users will believe that object is
ok, while in reality it happened to be be processed by accident). This
situation is however very unlikely. So in order to not introduce too
many log messages, that situation is skipped.

As long as errors of at least MINOR level on attributes exist,
MINOR message will be logged for corresponding object.