Move check for broken descriptors to higer level
We check for discrepancy with specification the earliest possible. While
we can't decide right away if 'attribute' flags are correct as it
depends on whether attribute belongs to object or template, we can do so
for 'object's and 'set's.
By doing so we "fixed" the flags early, thus partly obscuring
information.

Now all the checks are postponed until before the parsing affected items
themselves. No flags are overwritten.

Commit doesn't affect existing behavior.