Preserve platform-dependent variables in CircleCI
Dealing with parameters and variables is not very straightforward in
CircleCI.
1. There is no variable interpolation in CircleCI config
2. There are no modification operations that are applicable inside
config
3. Thus if variable value depends on parameter, but must modify it, it's
impossible to use default "environment" clause.

It is however possible to set up environment variables in powershell.
These variables will be accessible during the whole job.

(Other option to solve this problem is to create many parameters and
divide jobs to x32 and x64 ones. Unfortunately, that creates a lot of
duplication on the job level, which makes config even more confusing
than it is now. So it was decided against using this approach despite
its benefits.)