Assure exception handling model used in MSVC
From the [1]:
"The default exception unwinding code doesn't destroy automatic C++
objects outside of try blocks that go out of scope because of an
exception. Resource leaks and undefined behavior may result when a C++
exception is thrown."

At the moment we have 3 targets that might be affected: dlisio,
dlisio-extension and core. According to investigation, only
dlisio-extension actually requires exception handling. However, it is
possible that some errors simply are not catched by the test.

Hence take decision away from the user and assure MSV exception model
for all the targets.
Recommended model is /EHsc, which is therefore used.

[1]
https://docs.microsoft.com/en-us/cpp/build/reference/eh-exception-handling-model?view=msvc-160