Add mpark::variant as public dependency
The extension target was never installed properly by our build scripts.
If we are to do so, we would either need to install the vendored mpark
currently in external/mpark, or be open about mpark being a public
dependency. The latter means we do not bundle our own mpark and users
are given the freedom to use their own mpark version.