Add units to describe
While we don't have any code interface to present users with units, we
can at least print them in describe to make this information available
to researchers of individual files.

Only certain objects are updated.
It absolutely is possible to display units for everything, but it seems
to make little sense for objects such as Origin or Long-Name, where
values are expected to be strings.