Port the C-API of the dlisio target to a C++ API
With this we essentially give up on maintaining the dlisio target as a
pure C API. dlisio's current structure with a core C API and a C++
extension were never meant to be permanent - and the extension has grown
far beyond it's initial purpose.  The current division between the two
targets are neither sensible or particularly maintainable in the long
run. Thus the two targets will be merged together in one common C++ API.
Although a pure C API would be more portable - w.r.t binding other
languages, the decision fell on C++ as porting existing C++ code to C
was considered not to be worth the extra effort in the long run.

This commit only ports the C-API of the dlisio target into a C++-API.
The two targets will be merged in another commit. The implementation was
always the C++, and is untouched by this commit. The interface is vastly
kept as is, but with some overarching changes including:

 - The visible symbol DLISIO_API is removed
 - All public functions are namespaced under dl
 - The C-ish `dlis_`-prefix in all function signatures is removed.
 - All #define's are replaced with less global namespaced constexpr's

Additionally a new naming convention for type-related functions is
introduced:

 - dlis_sshort  -> dl::sshort_frombytes
 - dlis_sshorto -> dl::sshort_tobytes