Merge the extension into the dlisio-target
It was never the plan to maintain to separate targets for dlisio in the
way we do today. The extension target was a temporary way of letting us
use some c++ features for fast development. It has now grown to house
integral parts of dlisio.

This commit merges the two targes into one - essentially giving up on
maintaining a pure C API and acknowledging the extension as a vital part
of dlisio. It does not alter the current API's in any way. The C-API
from the original dlisio-target is keep intact. This may be subject to
change in the future