Remove fmtlib as submodule
The convenience of fmtlib as a submodule was lost when lfp was added as
a dependency. lfp also needs fmtlib and with the current setup they both
have their own copy of it. Although that is arguably ok, it takes away
any appeal of git submodules. This commit removes fmtlib as submodule
and requires it to be available through cmake's find_package().