Escape fseek error in findoffsets
As we previously attempted to surround in try/catch only cases which we
know could lead to exception, we missed a case where fseek throws errors
on reading and processing internal lfp header (rp66/tapeimage).

Remaining fseek entries which are now not surrounded in try-catch in
findoffsets are not expected to cause errors as they shouldn't read any
new headers. While some lfp-independent io errors are still possible,
they are not taken into consideration at the moment.