Move all tapemark related code to dlisio namespace
TapeImageFormat (TIF) is not specific to dlis, hence keeping it in the
top level namespace for re-use by the upcomming lis79 parser. Note that
the TIF support in dlisio is limited to whats needed in order to
identify a file as TapeImageFormat, such that this can be communicated
to lfp.