Initial commit for dlisio's LIS79 parser
This is an initial commit for dlisio's new Log Information Standard 79
(LIS79) parser. The parser is not feature-complete at this point - and
more testing and documentation is needed. However the fundamentals are
in place.

Like the dlis-parser the core is implemented in C/C++ with python
bindings on top. In the core the parser is namespaced under
dlisio::lis79 and in python as submodule dlisio.lis.

The diff is rather large, but splitting it up into stand-alone commits
that makes sense seams pointless.