Remove iodevice::poffset and iodevice::psize
poffset and psize was an attempt at supplying enough information
through the iodevice such that a new iodevice could be opened where
another ends. Which is essentially what we want do to when partitioning
a regular file into logical files.

However it turns out that due to variations of how tapemarks appear
between logical files it's not guaranteed that poffset and psize will
produce the correct tell for the next iodevice to be opened at. Without
that guarantee these functions loose their purpose.

It's now the caller, which have more information about the file as
whole, that is left to figure out where to open the next iodevice at.