Partition LIS LF's based on fixed records
This commit re-implements the logic for partition a file into logical
files. The old partitioning was too simplistic in that it was _only_
based on EOF and EOF tapemarks. In fact, it did not work at all for
files without tapemarks.

In this commit the partitioning logical is extended also include fixed
records and logical EOF records. The LIS79 specification defines that
the presence of Reel/Tape/File Header/Trailer Records mark the beginning
or end of a Reel, Tape or File, respectably.

This commit also touches the interface of the physical file as the
'Special LIS79 Logical Files' are no longer considered as logical files
when creating the physical_file. Rather, the Reel and Tape records
contained in these special logical files are extracted at load, and can
be accessed through the interface of logical_file.

Co-authored-by: Alena Chaikouskaya <alec@equinor.com>