Re-organise the sphinx documentation layout
With the introduction of a LIS parser the layout of our sphinx
documentation is in need of some restructuring. All dlis-specifics are
grouped under a new main section: Digital Log Interchange Standard, and
the source files are move into the subdirectory python/docs/dlis and
renamed appropriately.

This commit does not add any LIS documentation, it only makes room for
the LIS documentation to come.