Add new special cases load tests
Commit adds or updates a number of special-case test situations that
currently are failing.
Goal of the following commits would be to restructure our load functions
such a way that these cases are no longer special and are covered by
common code.

Also some error messages verifications are added. They are also
expected to be implemented.