Let stream::read return early when reading 0 bytes
This commit prevents dlisio from passing nullptr's as write-destination
when calling lfp_readinto, by checking the requested read-size in
stream::read.

Background on why dlisio would ever pass nullptr's to lfp_readinto:

Both lis::record and dlis::record uses std::vectors as buffers.
Populating these records with raw bytes is based on header information.
However, these methods do not have any special casing for emtpy records
and will call stream::read(buffer.data(), 0), which should be valid
call. Because the vector is empty, asking for a pointer to the
underlying data will return a nulltpr.

The most natural solution would be to give the responsibillity of not
passing a nullpr to the caller code. However, the amount of special
casing we need to dot around outways the benefit. Hence we implement the
simpler, but less correct approach of letting stream::read do an early
return if the requested number of bytes to read is 0.