Handle OSError's from core.openlis in lis.load
OSError's are emitted when core.openlis (lis::open) is asked to open a
new filehandle at and invalid position in the file. This typically would
happen if the file is truncated or have irregular padding.

This commits turns these OSError's into logging.errors, which means that
any LF's that where succesfully indexed prior to the error can be
returned by load.