Upgrade latest pip version in CI pipelines
This commit makes sure we consistently use the latest version of pip in
our CI pipelines. This is basically to ensure we use a pip version with
the latest dependency resolver [1]. The new, stricter rules of the
dependency resolver enforces that dependencies are correctly upgraded to
their minimum required version.

[1] https://pip.pypa.io/en/stable/user_guide/#resolver-changes-2020