Check for multi-dimensional curves in lis.curves
dlisio does not support multi-dimension curves at this point. This
commit adds an check for the presence of such curves before trying to
read the actual curve data. The lack of this check was a bug and trying
to read such Data Format Specification records would fail or worse -
return incorrect data.