Update LIS layout tests
Layouts are based on the ones present in the real files and a bit from
our knowledge for dlis.

Some existing tests are removed, mainly the ones with no indication that
such files exist in reality.
It also includes files ending on just 1 TM(1) as this configuration has
never been seen in dlis or in currently investigated .lis files. While
it's possible that they exist, it would make more sense to lock the
behavior only once this file is discovered.